refactor(client): fix theme error, move service-worker to directory
