+++ packages/client/src/serviceWorker/service-worker.ts
