+++ packages/client/src/serviceWorker/serviceWorkerRegistration.ts
@@ -10,7 +10,7 @@
 // To learn more about the benefits of this model and instructions on how to
 // opt-in, read https://cra.link/PWA
 
-import { Environment } from './constants';
+import { Environment } from '../constants';
 
 const isLocalhost = Boolean(
   window.location.hostname === 'localhost' ||
@@ -37,7 +37,7 @@ export function register(config?: Config) {
     }
 
     window.addEventListener('load', () => {
-      const swUrl = `${process.env.PUBLIC_URL}/service-worker.js`;
+      const swUrl = `${process.env.PUBLIC_URL}/serviceWorker/service-worker.js`;
 
       if (isLocalhost) {
         // This is running on localhost. Let's check if a service worker still exists or not.
